+++ src/sync-plugin-key.ts
@@ -1,0 +1,23 @@
+import type { ContainerID, LoroDoc, Subscription } from "loro-crdt";
+import { PluginKey } from "prosemirror-state";
+import type { EditorView } from "prosemirror-view";
+import type { LoroDocType, LoroNodeMapping } from "./lib";
+
+export const loroSyncPluginKey = new PluginKey<LoroSyncPluginState>(
+  "loro-sync",
+);
+
+export interface LoroSyncPluginProps {
+  doc: LoroDocType;
+  mapping?: LoroNodeMapping;
+  containerId?: ContainerID;
+}
+
+export interface LoroSyncPluginState extends LoroSyncPluginProps {
+  changedBy: "local" | "import" | "checkout";
+  mapping: LoroNodeMapping;
+  snapshot?: LoroDoc | null;
+  view?: EditorView;
+  containerId?: ContainerID;
+  docSubscription?: Subscription | null;
+}
